requirements.txt
tensorflow>=2.10
numpy
pandas
matplotlib
seaborn
flask
Pillow
scikit-learn
